Fix placeholder formatting in Thread Primer (#19)
We had to adjust our openthread.io Copybara import to account for some
new code block variations, which ended up conflicting with this one
existing example in the Thread Primer. So I've rewritten them here so
that it's no longer an issue.

This is not a problem in the Chinese version of the Thread Primer
since it's formatted slightly different (it doesn't have a period
directly after the ending code tag), so that is unchanged.